+++ src/com/solinia/solinia/Managers/UniversalTemporarySoliniaLootTable.java
@@ -3,10 +3,10 @@ package com.solinia.solinia.Managers;
 import com.solinia.solinia.Interfaces.ISoliniaLootTable;
 
 public class UniversalTemporarySoliniaLootTable {
-	public int distance;
+	public long distance;
 	public ISoliniaLootTable soliniaLootTable;
 	
-	public UniversalTemporarySoliniaLootTable(ISoliniaLootTable soliniaLootTable, int distance)
+	public UniversalTemporarySoliniaLootTable(ISoliniaLootTable soliniaLootTable, long distance)
 	{
 		this.distance = distance;
 		this.soliniaLootTable = soliniaLootTable;
